test: add sudotest, update tests to pass as sudo

+++ test/common-tap.js
@@ -22,6 +22,7 @@ if (!global.setImmediate || !require('timers').setImmediate) {
 }
 
 var spawn = require('child_process').spawn
+const spawnSync = require('child_process').spawnSync
 var path = require('path')
 
 // provide a working dir unique to each test
@@ -30,9 +31,27 @@ exports.pkg = path.resolve(path.dirname(main), path.basename(main, '.js'))
 const mkdirp = require('mkdirp')
 const rimraf = require('rimraf')
 mkdirp.sync(exports.pkg)
+
 require('tap').teardown(() => {
   try {
-    rimraf.sync(exports.pkg)
+    const isRoot = process.env.getuid && process.env.getuid() === 0
+    const isSudo = isRoot && process.env.SUDO_UID
+    if (isSudo) {
+      // running tests as sudo.  ensure we didn't leave any root-owned
+      // files in the cache by mistake.
+      const args = [ npm_config_cache, '-uid', '0' ]
+      const found = spawnSync('find', args)
+      const output = found.stdout.toString()
+      if (output.length) {
+        const er = new Error('Root-owned files left in cache!')
+        er.files = output.trim().split('\n')
+        throw er
+      }
+    }
+    if (!process.env.NO_TEST_CLEANUP) {
+      rimraf.sync(exports.pkg)
+      rimraf.sync(npm_config_cache)
+    }
   } catch (e) {
     if (process.platform !== 'win32') {
       throw e
@@ -60,6 +79,7 @@ ourenv.npm_config_metrics = 'false'
 ourenv.npm_config_audit = 'false'
 
 var npm_config_cache = path.resolve(__dirname, 'npm_cache_' + testId)
+ourenv.npm_config_unsafe_perm = 'true'
 ourenv.npm_config_cache = exports.npm_config_cache = npm_config_cache
 ourenv.npm_config_userconfig = exports.npm_config_userconfig = configCommon.userconfig
 ourenv.npm_config_globalconfig = exports.npm_config_globalconfig = configCommon.globalconfig
@@ -95,6 +115,9 @@ exports.npm = function (cmd, opts, cb) {
   if (opts.env._storage) opts.env = Object.assign({}, opts.env._storage)
   if (!opts.env.npm_config_cache) {
     opts.env.npm_config_cache = npm_config_cache
+  }
+  if (!opts.env.npm_config_unsafe_perm) {
+    opts.env.npm_config_unsafe_perm = 'true'
   }
   if (!opts.env.npm_config_send_metrics) {
     opts.env.npm_config_send_metrics = 'false'
